AND-27 Get all events call added

+++ presentation/src/main/java/com/activity_sync/presentation/models/Level.java
@@ -21,7 +21,7 @@ public class Level
 
     }
 
-    public double getId()
+    public int getId()
     {
         return id;
     }
